[IMP] sale_timesheet: store remaining_hours field in SOL model
When the remaining_hours field is stored, we can search the last
SOL and the remaining hours on SO is greater than 0 for a customer with
only a search in SOL model. With this changes, we increase the
performance because before this commit, we search and retrieve a list of
SOLs and then we filter this list to find the last SOL with remaining
hours on SO is greater than 0.

This commit stores this field in sale.order.line model and we update the
_get_last_sol_of_customer method in project.task.

task-2424382